requirements.txt
pystray
praw
Pillow
win10toast-click
